move and rename AndOrStubsUtil to CourseTestUtil
