+++ src/test/java/seedu/jarvis/commons/util/andor/AndNodeTest.java
@@ -3,8 +3,8 @@ package seedu.jarvis.commons.util.andor;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.jarvis.commons.util.andor.AndOrStubsUtil.CHILDREN;
-import static seedu.jarvis.commons.util.andor.AndOrStubsUtil.CourseStub;
+import static seedu.jarvis.testutil.course.CourseTestUtil.CHILDREN;
+import static seedu.jarvis.testutil.course.CourseTestUtil.CourseStub;
 
 import java.util.List;
