+++ src/test/java/seedu/jarvis/commons/util/andor/LeafNodeTest.java
@@ -3,7 +3,7 @@ package seedu.jarvis.commons.util.andor;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.jarvis.commons.util.andor.AndOrStubsUtil.CourseStub;
+import static seedu.jarvis.testutil.course.CourseTestUtil.CourseStub;
 
 import java.util.List;
